change: Stat avg to float

+++ src/main/scala/mpt/lab/one/stat/Stat.scala
@@ -36,6 +36,6 @@ class Stat {
   }
 
   /** Среднее число операций для элемента */
-  def avg() = allElementsCounter.toDouble / elementsCount
+  def avg() = allElementsCounter.toFloat / elementsCount
 }
